Rid warnings.

+++ Engine/source/vkChunkAllocator.cpp
@@ -15,17 +15,17 @@ std::unique_ptr<Chunk> ChunkAllocator::Allocate(VkDeviceSize size, int memoryTyp
 
 VkDeviceSize ChunkAllocator::NextPowerOfTwo(VkDeviceSize size)
 {
-	VkDeviceSize power = static_cast<VkDeviceSize>(std::log2l(size) + 1);
-	return static_cast<VkDeviceSize>(1 << power);
+	VkDeviceSize power = static_cast<VkDeviceSize>(std::log2l(static_cast<long double>(size)) + 1);
+	return static_cast<VkDeviceSize>(static_cast<uint64_t>(1) << power);
 }
 
 bool ChunkAllocator::IsPowerOfTwo(VkDeviceSize size)
 {
 	VkDeviceSize mask = 0;
-	VkDeviceSize power = static_cast<VkDeviceSize>(std::log2l(size));
+	VkDeviceSize power = static_cast<VkDeviceSize>(std::log2l(static_cast<long double>(size)));
 
 	for (VkDeviceSize i = 0; i < power; ++i)
-		mask += static_cast<VkDeviceSize>(1 << i);
+		mask += static_cast<VkDeviceSize>(static_cast<uint64_t>(1) << i);
 
 	return !(size & mask);
 }
